Added TMap, TSet and some other UE stuff Added new API functions Updated FString, TArray

+++ version/Core/Public/API/Base.h
@@ -139,10 +139,10 @@ struct UStaticMesh;
 
 struct BitField
 {
-	DWORD64 offset;
-	unsigned int bit_position;
-	unsigned int num_bits;
-	unsigned int length; //in bytes
+	LONG offset;
+	DWORD bit_position;
+	ULONGLONG num_bits;
+	ULONGLONG length; //in bytes
 };
 
 // Address helpers
